Fixed new function isTypeOf
