+++ README.md
@@ -91,7 +91,7 @@ const count = stringzy.analyze.wordCount('Hello world'); // 2
 - [isDate](#isdate) - Checks if a string is a valid date
 - [isEmpty](#isempty) - Checks if a string is empty or contains only whitespace
 - [isSlug](#isslug) - Checks if a string is a valid slug
-- [isType](#istype) - Checks if a file or URL has a valid extension for a given type
+- [isTypeOf](#istypeof) - Checks if a file or URL has a valid extension for a given type
 - [isIPv4](#isipv4) - Checks if a string is a valid IPv4 address
 - [isHexColor](#ishexcolor) - Checks if the input string is a valid hex color
 
@@ -515,7 +515,7 @@ isSlug("hello_world");         // false (underscore not allowed)
 |-----------|------|---------|-------------|
 | text | string | required | The input string to validate as slug |
 
-#### <a id="istype"></a>`isType(input, type)`
+#### <a id="istypeof"></a>`isTypeOf(input, type)`
 
 Checks if a file or URL has a valid extension for a given type
 
@@ -530,7 +530,7 @@ isType("video.mp4", "document");                 // false
 
 | Parameter | Type | Default | Description |
 |-----------|------|---------|-------------|
-| input | string, object | required | The file name, URL string, or object with .name |
+| input | string | required | The file name, URL string, or object with .name |
 | input | string | required | The file type category to validate (image, video, audio, document, archive) |
